重大更新到2.0.0

+++ kinit-api/apps/vadmin/auth/utils/validation/login.py
@@ -7,13 +7,12 @@
 # @desc           : 登录验证装饰器
 
 from fastapi import Request
-from pydantic import BaseModel, validator, field_validator
+from pydantic import BaseModel, field_validator
 from sqlalchemy.ext.asyncio import AsyncSession
-from application.settings import DEFAULT_AUTH_ERROR_MAX_NUMBER, DEMO
+from application.settings import DEFAULT_AUTH_ERROR_MAX_NUMBER, DEMO, REDIS_DB_ENABLE
 from apps.vadmin.auth import crud, schemas
 from core.database import redis_getter
 from core.validator import vali_telephone
-from typing import Optional
 from utils.count import Count
 
 
@@ -64,12 +63,15 @@ class LoginValidation:
 
         result = await self.func(self, data=data, user=user, request=request)
 
-        count_key = f"{data.telephone}_password_auth" if data.method == '0' else f"{data.telephone}_sms_auth"
-        count = Count(redis_getter(request), count_key)
+        if REDIS_DB_ENABLE:
+            count_key = f"{data.telephone}_password_auth" if data.method == '0' else f"{data.telephone}_sms_auth"
+            count = Count(redis_getter(request), count_key)
+        else:
+            count = None
 
         if not result.status:
             self.result.msg = result.msg
-            if not DEMO:
+            if not DEMO and count:
                 number = await count.add(ex=86400)
                 if number >= DEFAULT_AUTH_ERROR_MAX_NUMBER:
                     await count.reset()
@@ -81,10 +83,10 @@ class LoginValidation:
         elif data.platform in ["0", "1"] and not user.is_staff:
             self.result.msg = "此手机号无权限！"
         else:
-            if not DEMO:
+            if not DEMO and count:
                 await count.delete()
             self.result.msg = "OK"
             self.result.status = True
             self.result.user = schemas.UserSimpleOut.model_validate(user)
-            await user.update_login_info(db, request.client.host)
+            await crud.UserDal(db).update_login_info(user, request.client.host)
         return self.result
